requirements.txt
streamlit==0.89.0
pytesseract==0.3.8
pillow==8.4.0
spacy==3.1.3
pdf2image==1.16.0
python-docx==0.8.11
